C:\Users\zainu\Desktop\курс\proekt2\Ozhaev Zaynudin Sharik5.cpp -o C:\Users\zainu\Desktop\курс\proekt2\Ozhaev Zaynudin Sharik5.exe  -Wshadow -Winit-self -Wredundant-decls -Wcast-align -Wundef -Wfloat-equal -Winline -Wunreachable-code -Wmissing-declarations -Wmissing-include-dirs -Wswitch-enum -Wswitch-default -Weffc++ -Wmain -Wextra -Wall -g -pipe -fexceptions -Wcast-qual -Wconversion -Wctor-dtor-privacy -Wempty-body -Wformat-security -Wformat=2 -Wignored-qualifiers -Wlogical-op -Wmissing-field-initializers -Wnon-virtual-dtor -Woverloaded-virtual -Wpointer-arith -Wsign-promo -Wstack-usage=8192 -Wstrict-aliasing -Wstrict-null-sentinel -Wtype-limits -Wwrite-strings -D_DEBUG -D_EJUDGE_CLIENT_SIDE -D_EJC

+++ Ozhaev Zaynudin Sharik5.cpp
@@ -231,8 +231,8 @@ void CollisionBall (Ball* ball_1, Ball* ball_2, int dt)
     printf ("Coordin do(): Dy  = %d\n", Dy);
     printf ("Coordin do(): d  = %f\n", d);
     //txSleep (5000);
-    float sin = Dx/d; // sin угла треугольника
-    float cos = Dy/d; // cos угла треугольника
+    double sin = Dx/d; // sin угла треугольника
+    double cos = Dy/d; // cos угла треугольника
 
     if (d < (*ball_1) .r + (*ball_2) .r) //проверка столкновения
         {
